requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.3.1
drf-yasg==1.21.7
psycopg2-binary==2.9.10
django-filter==24.1
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
Pillow==10.3.0
